Remove one unused HTML id in tile-icons.html
IDs surface to user-space: they can be used as URL fragments.
Thou shalt not break user-space. Keeping what you expose to a minimum helps
to respect that commandment.

Removing or renaming an ID leads, potentially, to broken links (strictly
speaking — but if pragmatically a broken fragment may just be a minor
annoyance, it is still an annoyance.)

In my book, if someone breaks a digital link of yours, you are entitled
to break one of his/her finger IRL.

Now I _am_ potentially breaking some links with this commit. But the site is
very new and very little known at the moment, so I’ll take the risk of someone
going after a finger of mine ;-)