Add a custom HTTP 404 page
Netlify provides a nice 404 page by default, but it uses the Google
Fonts service. So rendering the page blocks on the Font CSS request,
and, more importantly, there is tracking by Google. I don’t like that.

Moreover, most 404 are going to occur because of script kiddie bots
that crawl the Web searching for some WordPress/whatever to hack. Hence
it’s good to keep the 404 page weight to a minimum (so as to not waste
the Internet global bandwidth), but Netlify default page is not
especially light.