Generate & handle the required 50px² Store icon
The Microsoft Store does not accept (& scale) a larger existing icon.

Up to this commit, generated packages were rejected, in a late stage,
during the submission process to the Store!
(I never actually went to the end of the process until now 😑😅)

Thanks to the gentleman who reported me this issue by email.