Add missing ‘since’ JSDoc tag in appx-manifest.js
For conformity with the other JS files, and because it can be useful
to know when a piece of code was written.

Note that the ‘since’ tag is meant to hold a version identifier.
See http://usejsdoc.org/tags-since.html
I consider a date to be a valid version id, in this case.