Remove CSS ‘scroll-behavior’ in appx-manifest.html
It’s redundant with the Scroll{IntoView}Options used in the JS code.
See https://drafts.csswg.org/cssom-view/#dictdef-scrollintoviewoptions
See https://developer.mozilla.org/docs/Web/CSS/scroll-behavior

Why not do it in CSS? Because it’s hard to make that work properly,
at least in this case.

I try setting “scroll-behavior: smooth” on #output, the containing
<form>, and the containing <main> element: no luck, the scrolling is
brutal, on Chrome & Firefox. Adding “overflow-y: {auto,scroll}” didn’t
change anything. Messing with the {Scroll,Focus}Options in the JS code
didn’t help either.

For Chrome, “scroll-behavior: smooth” has to be set on the html element
to be effective. (On Firefox, setting it on the body element is enough.)

As a result, on Chrome, the scroll will also be smooth when focusing the
status element, once the user saves the manifest in his/her package.
But commit 53116132deb0911072075c0783a25b32e00cf805 was all about
“ensur[ing] status animation is in-view, focus brutally”! We could
probably change again the way we focus the status element so as to make
it explicitely brutal, but why bother when we can just use JS here?