Tweak [size] of publisher id <input>, for Chrome
My Microsoft publisher id (“CN=…”) is 39 characters long.

On Chromium 71 Linux, [size=48] was too small: the last char got
truncated. On Firefox 65 Linux, [size=48] was already too wide.

(I’m assuming that most publishers id share the same format/length.)

-----

Reduced test case: “<!doctype html><input size=10 lang=en>”.

Input “0123456789”. On Chrome: you see all chars, and there’s no extra
space on the right. On Firefox, there’s ~40% extra space on the right.

But now input “WWWWWWWWWW” (10 ‘W’, which is the widest latin letter
in most fonts). On Chrome: you see only 7 of them (and a tiny bit of
the 8th). On Firefox, you can see the 10 (though the last one is
slightly truncated).

It seems Chrome use the equivalent of CSS “width: 10ch” (1 ch is
equal to the width of the ‘0’ char). Since the default font is not
monospace, but because the digits actually are of the same width, you
get the behavior just described.

What happens if you set “font-family: monospace” on the <input>?
On Firefox, you get perfect sizing. On Chrome, you get extra space on
the right this time, enough for one full character!

The spec says: “The size attribute gives the number of characters that,
in a visual rendering, the user agent is to allow the user to see while
editing the element’s value.”
https://www.w3.org/TR/html52/sec-forms.html#the-size-attribute

Chrome is in violation of the spec. [Hence the fact that my 39-char id
cannot fit into an <input size=48>!]

Anyway, spec-compliance per se does not matter much. The real issue here
is the huge difference in size, by default, between Firefox & Chrome (I
did not try with other browsers). This problem exists since at least 10 
years; that’s when I started doing web dev seriously, and I remember
already encountering this issue back then.

Standardization between browser vendors on a sane default size (range)
for <input> is much needed. It’s not like <input> is rarely used. Given
that such work would not be “exciting” or “fun” though, this is probably
never gonna happen. At least remove [size] from the HTML spec, then: in
its current state, it is deceiving as hell.