Make it possible to “import” a web app manifest
…to pre-fill the appx manifest form.

The core of that job is quite trivial. The trickiest part is dealing
with colors. That’s well-explained in the comments of appx-manifest.js

This commit leaves the codebase in a sorry, ugly state. Some JS code
cleaning will have to happen, and, more importantly, something will
have to be done to keep the CSS selectors sane. The HTML elements will
probably get more classy.  This work will come next.

I would have preferred to use a <dialog> for making the user input
his/her Web App Manifest, but few browsers support <dialog>, and a
polyfill would probably be as much code as all of appx-manifest.js
(you’ve got to trap focus, aria-hide everything but the <dialog>, &c.)

The user is required to copy+paste his/her Web App Manifest manually.
HTTP-fetching it automatically would only be possible if the manifest
were served with proper CORS headers, and I doubt that’s commonly the
case, sadly.