Fix DOM state inconsistency when resetting form
…in appx-manifest.js, regarding the read/write state of the tile icon
path <input>s.

Suppose the user unchecks the “Use same icon paths than the tile icons
generator” checkbox. This makes the icon path <input>s be read-write.
Now the user clicks the “Reset the form” button. This checks back the
checkbox, but it previously did not make the path <input>s revert to
being read-only (and that was a minor defect).

This makes the point for using an immediate-mode user interface / a
virtual DOM. See https://twitter.com/sebmarkbage/status/4573473271717890
and https://caseymuratori.com/blog_0001
But in our case, the UI is (still) too simple to justify going that way.